move deserializeTxList
